[SDK][DRIVERS][WIN32SS] Fix the GCC build and clear its warnings
fix: close the pci.h include guard after the memory-descriptor helpers
fix: use MAXULONG in dxgkrnl gpuva, limits.h is not in scope
fix: link rosget against cppstl so the STL headers are visible
fix: keep one math.h in glmark2 C++ TUs under libstdc++
fix: place the WDF driver-entry markers with DATA_SEG
fix: drop the ignored allocate attribute from the WDF type-info macro
fix: honour _KERNELBASE_ in the WINBASEAPI import guard
fix: emit the d3d10/d3d11 helper diagnostics for C++ only
fix: let the disabled dxgkrnl trace macros consume their arguments
fix: return from the umpnpmgr device-event worker
fix: brace the SID authority and KS property-set initialisers
fix: drop dead locals, helpers and a misleading menu clamp

+++ base/applications/cmdutils/rosget/sqlite_index.cpp
@@ -250,7 +250,7 @@ Status SQLiteIndex::ScanTablePage(std::uint32_t pageNumber, const RowCallback &c
             return Status::Fail(ERROR_BAD_FORMAT, "WinGet index B-tree cell points outside its page");
         if (interior)
         {
-            if (cellOffset + 4 > pageSize_)
+            if (static_cast<std::uint32_t>(cellOffset) + 4 > pageSize_)
                 return Status::Fail(ERROR_BAD_FORMAT, "WinGet index interior cell is truncated");
             Status status = ScanTablePage(ReadBe32(page + cellOffset), callback, visited);
             if (!status)
@@ -323,7 +323,7 @@ Status SQLiteIndex::ScanIndexPage(std::uint32_t pageNumber, const RowCallback &c
         std::size_t cursor = cellOffset;
         if (interior)
         {
-            if (cellOffset + 4 > pageSize_)
+            if (static_cast<std::uint32_t>(cellOffset) + 4 > pageSize_)
                 return Status::Fail(ERROR_BAD_FORMAT, "WinGet index interior cell is truncated");
             Status status = ScanIndexPage(ReadBe32(page + cursor), callback, visited);
             if (!status)
